Separate Terminal from Logger, use Terminal as default transport
